+++ Top/csound.c
@@ -236,18 +236,18 @@ static void create_opcode_table(CSOUND *csound)
 }
 
 
-static int32_t sndfileWrite(CSOUND *csound, void *h, MYFLT *p, int32_t frames){
+static int64_t sndfileWrite(CSOUND *csound, void *h, MYFLT *p, int64_t frames){
   IGN(csound);
-  return (int32_t) sflib_writef_MYFLT(h,p,frames); 
+  return sflib_writef_MYFLT(h,p,frames); 
 }
 
-static int32_t sndfileRead(CSOUND *csound, void *h, MYFLT *p, int32_t frames){
+static int64_t sndfileRead(CSOUND *csound, void *h, MYFLT *p, int64_t frames){
   IGN(csound);
-  return (int32_t) sflib_readf_MYFLT(h,p,frames); 
+  return sflib_readf_MYFLT(h,p,frames); 
 }
 
-static int32_t sndfileSeek(CSOUND *csound, void *h, int32_t frames, int32_t whence){
-  return (int32_t) sflib_seek(h, frames, whence);
+static int64_t sndfileSeek(CSOUND *csound, void *h, int64_t frames, int32_t whence){
+  return sflib_seek(h, frames, whence);
 }
 
 #define MAX_MODULES 64
